feat: add push notification dispatching for event processing
